+++ lua/plugins/carbon-now.lua
@@ -1,7 +1,6 @@
 ---@type LazyPluginSpec
 return {
    "ellisonleao/carbon-now.nvim",
-   lazy = true,
    cmd = { "CarbonNow" },
    opts = {
       options = {