+++ lua/plugins/nvim-autopairs.lua
@@ -1,6 +1,7 @@
 ---@type LazyPluginSpec
 return {
    "windwp/nvim-autopairs",
+   event = "InsertEnter",
    opts = {
       fast_wrap = {},
       disable_filetype = { "TelescopePrompt", "vim" },