+++ lua/plugins/nvim-surround.lua
@@ -3,6 +3,6 @@ return {
    "kylechui/nvim-surround",
    -- Use for stability; omit to use `main` branch for the latest features
    version = "*",
-   event = "VeryLazy",
+   event = "InsertEnter",
    config = true,
 }