+++ lua/plugins/outline.lua
@@ -1,7 +1,6 @@
 ---@type LazyPluginSpec
 return {
    "hedyhli/outline.nvim",
-   lazy = true,
    cmd = { "Outline", "OutlineOpen" },
    keys = require("core.keymaps").outline.keys,
    opts = {