+++ lua/plugins/typst-preview.lua
@@ -1,6 +1,6 @@
 return {
    "chomosuke/typst-preview.nvim",
-   ft = "typst",
    version = "1.*",
+   ft = "typst",
    opts = {},
 }