+++ lua/plugins/undotree.lua
@@ -1,7 +1,0 @@
----@type LazyPluginSpec
-return {
-   "mbbill/undotree",
-   lazy = true,
-   cmd = "UndotreeToggle",
-   keys = require("core.keymaps").undotree.keys,
-}