+++ lua/plugins/vim-floaterm.lua
@@ -1,4 +1,5 @@
 ---@type LazyPluginSpec
 return {
    "voldikss/vim-floaterm",
+   cmd = { "FloatermNew", "FloatermToggle", "FloatermPrev", "FloatermNext" },
 }